% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_CellCounting.R
\name{ArchR_cell_counting}
\alias{ArchR_cell_counting}
\title{ArchR function: Plot table of cell counts from different groups}
\usage{
ArchR_cell_counting(
  ArchR = ArchR,
  group1 = "clusters",
  group2 = "stage",
  print_table = TRUE,
  scHelper_cell_type_order = scHelper_cell_type_order
)
}
\arguments{
\item{ArchR}{ArchR object}

\item{group1}{first column to split cells by}

\item{group2}{second column to split cells by}

\item{print_table}{boolean to print table or return it}

\item{scHelper_cell_type_order}{order in which to return cell types}
}
\value{
table of cell counts, or plots table directly
}
\description{
This function counts how many cells from each cluster/sample are assigned the same label/cluster
}
