% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_CellCountsHeatmap.R
\name{ArchR_cell_counts_heatmap}
\alias{ArchR_cell_counts_heatmap}
\title{ArchR function: Plot heatmap of cell counts}
\usage{
ArchR_cell_counts_heatmap(
  ArchR = ArchR,
  group1 = "scHelper_cell_type",
  group2 = "clusters"
)
}
\arguments{
\item{ArchR}{ArchR object}

\item{group1}{first column to split cells by}

\item{group2}{second column to split cells by}
}
\value{
plot generated by pheatmap
}
\description{
This function plots a heatmap of cell counts split across two groups
}
