% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_CellCountsPiecharts.R
\name{ArchR_cell_counts_piecharts}
\alias{ArchR_cell_counts_piecharts}
\title{ArchR function: Plot table of cell counts from different groups}
\usage{
ArchR_cell_counts_piecharts(counts, cols, scale = FALSE)
}
\arguments{
\item{counts}{???}

\item{cols}{???}

\item{scale}{whether to scale by total number of cells}
}
\value{
grid of piecharts
}
\description{
This function makes piecharts/barcharts showing contribution of cell groups to other cell groups
}
