% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_ExtractMeansFromSe.R
\name{ArchR_extract_means_from_se}
\alias{ArchR_extract_means_from_se}
\title{ArchR function: Extract mean matrix from se object}
\usage{
ArchR_extract_means_from_se(seMarker, Log2norm = TRUE, scaleTo = 10^4)
}
\arguments{
\item{seMarker}{the se object created \code{ArchR::getMarkerFeatures}}

\item{Log2norm}{boolean to log2 normalize the matrix + scale for cluster size}

\item{scaleTo}{scale factor to multiply by}
}
\value{
means matrix for plotting
}
\description{
This function takes seMarker and extracts the mean matrix for heatmap plotting
}
