% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_ExtractIds.R
\name{ArchR_ExtractIds}
\alias{ArchR_ExtractIds}
\title{ArchR function: Extract IDs for plotting}
\usage{
ArchR_ExtractIds(
  seMarker,
  cutOff = "FDR <= 1 & Log2FC >= 0",
  top_n = TRUE,
  n = 10,
  group_name = "clusters"
)
}
\arguments{
\item{seMarker}{the se object created \code{ArchR::getMarkerFeatures}}

\item{cutOff}{string to filter features, eg by FDR or Log2FC}

\item{top_n}{boolean to extract top n features}

\item{n}{if top_n is TRUE, extract top n features}

\item{group_name}{name of grouping to be used when extracting top n features}
}
\value{
means matrix for plotting
}
\description{
This function extracts the IDs to be plotted, either by cut off or cut off + top n features
}
