% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_FeaturePlotGrid.R
\name{ArchR_FeaturePlotGrid}
\alias{ArchR_FeaturePlotGrid}
\title{ArchR function: Plots grid of featureplots}
\usage{
ArchR_FeaturePlotGrid(ArchR, matrix = "GeneScoreMatrix", feature_list)
}
\arguments{
\item{ArchR}{ArchR object}

\item{matrix}{which matrix to plot (default = "GeneScoreMatrix")}

\item{feature_list}{which features to plot}
}
\value{
grid of featureplots
}
\description{
This function makes a grid of featureplots
}
