% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_PlotMarkerHeatmap.R
\name{ArchR_PlotMarkerHeatmap}
\alias{ArchR_PlotMarkerHeatmap}
\title{ArchR function: Plots marker heatmap}
\usage{
ArchR_PlotMarkerHeatmap(
  mat,
  pal = viridis::magma(100),
  labelRows = FALSE,
  fontSizeRows = 12,
  labelCols = TRUE,
  fontSizeCols = 12,
  cluster_columns = TRUE,
  cluster_rows = TRUE
)
}
\arguments{
\item{mat}{data matrix, created by: Log2norm(extract_means_from_se(seMarker))}

\item{pal}{color palatte for heatmap cells}

\item{labelRows}{boolean, whether to label rows}

\item{fontSizeRows}{font size for row labels}

\item{labelCols}{boolean, whether to label cols}

\item{fontSizeCols}{font size for col labels}

\item{cluster_columns}{boolean, whether to cluster cols}

\item{cluster_rows}{boolean, whether to cluster rows}
}
\value{
Heatmap of marker features
}
\description{
This function takes matrix from seMarker and plots heatmap
}
