% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_pseudoreplicate_counts.R
\name{ArchR_PseudoreplicateCounts}
\alias{ArchR_PseudoreplicateCounts}
\title{ArchR function: Plot table of cell counts in pseudoreplicates}
\usage{
ArchR_PseudoreplicateCounts(
  ArchR = ArchR,
  pseudo_replicates,
  group_by = "Sample"
)
}
\arguments{
\item{ArchR}{ArchR object}

\item{pseudo_replicates}{object created by Archr::addGroupCoverages}

\item{group_by}{ArchR metadata to group by (e.g. "Sample")}
}
\value{
table of cell counts for each pseudoreplicate across groups
}
\description{
This function prints a table of how many cells are in each pseudoreplicate and how many samples and clusters are they made of
}
