% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_SplitbyStage.R
\name{ArchR_SplitbyStage}
\alias{ArchR_SplitbyStage}
\title{ArchR function: Split by stage}
\usage{
ArchR_SplitbyStage(ArchR)
}
\arguments{
\item{ArchR}{ArchR object}
}
\value{
list of ArchR objects, split by stage
}
\description{
This function splits ArchR object by stage
}
