% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageGeneModules.R
\name{AverageGeneModules}
\alias{AverageGeneModules}
\title{AverageGeneModules}
\usage{
AverageGeneModules(seurat_obj, gene_list)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene_list}{Named list of genes passed to seurat::AddModuleScore}
}
\value{
seurat_obj
}
\description{
Calculate average expression for groups of genes across Seurat clusters and add to seurat_obj@meta.data.
}
