% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.R
\name{BoxPlot}
\alias{BoxPlot}
\title{Box Plot}
\usage{
BoxPlot(data, y_col, group_by, y_lab, x_lab)
}
\arguments{
\item{data}{df containing sample metadata (i.e. seurat@meta.data)}

\item{y_col}{dat column to place on y axis}

\item{group_by}{dat column to group cells by on x axis}

\item{y_lab}{label for y axis}

\item{x_lab}{label for x axis}
}
\value{
ggplot object
}
\description{
This function generates a box plot from a single cell metadata dataframe
}
