% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellLabelPieCharts.R
\name{CellLabelPieCharts}
\alias{CellLabelPieCharts}
\title{Cell Labels Pie Charts}
\usage{
CellLabelPieCharts(counts, cols, scale = FALSE)
}
\arguments{
\item{counts}{df containing cell labels as columns and samples as rows}

\item{cols}{vector of colours to use for each cell label}

\item{scale}{boolean value to specify whether to scale the data by the number of cells in each sample}
}
\value{
ggplot object
}
\description{
This function generates a pie chart for each cell label showing the proportion of cells in each sample
}
