% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterClassification.R
\name{ClusterClassification}
\alias{ClusterClassification}
\title{ClusterClassification}
\usage{
ClusterClassification(
  seurat_obj,
  group_by = "seurat_clusters",
  metrics,
  quantile,
  annotation_name = "scHelper_cell_type"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{group_by}{column to group cells by}

\item{quantile}{Percent quantile to plot on bar plots. Must be a single value representing the upper percent quantile}

\item{annotation_name}{Name of column to be annotated with corresponding cell types}

\item{cell_types}{column names for cell type module scores found in seurat_obj@meta.data}
}
\value{
array of clusters which are outliers
}
\description{
This function retrieves cluster classifications using cell module score columns in seurat object. Clusters are classiffied as a given cell type if the median for a given cell type falls outside the provided quantile.
}
