% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDimplot.R
\name{ClusterDimplot}
\alias{ClusterDimplot}
\title{Cluster Dimplot}
\usage{
ClusterDimplot(
  seurat_object,
  group_by = "seurat_clusters",
  clusters,
  cluster_col = "seurat_clusters",
  pt.size = 0.2,
  xlim = c(-15, 15),
  ylim = c(-15, 15),
  nrow = 1,
  shuffle = TRUE,
  seed = 123
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{group_by}{seurat_object@meta.data column to group cells by in DimPlots}

\item{clusters}{array of clusters to subset and plot}

\item{cluster_col}{seurat_object@meta.data column for which clusters are found in}

\item{nrow}{number of rows to pass to grid.arrange}

\item{shuffle}{randomly plot cells in order to prevent hiding certain populations when overplotting}

\item{seed}{set seed for plotting consistency when shuffle is TRUE}
}
\value{
plot generated by grid arrange
}
\description{
This function plots UMAPs for selected clusters of interest
}
