% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustRes.R
\name{ClustRes}
\alias{ClustRes}
\title{Plot multiple cluster resolutions}
\usage{
ClustRes(seurat_object, by = 0.1, starting_res = 0, prefix = "RNA_snn_res.")
}
\arguments{
\item{seurat_object}{Seurat object}

\item{by}{value by which resolution is incremented}

\item{starting_res}{lowest resolution to cluster cells by}

\item{prefix}{prefix used to select metadata columns with cluster identities following clustering}
}
\value{
plot generated by grid extra
}
\description{
This function plots UMAPs for a series of nine cluster resolutions, alongside a clustree which reveals the relationship between clusters at different resolutions
}
