% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCellTypeAnnotation.R
\name{CreateCellTypeAnnotation}
\alias{CreateCellTypeAnnotation}
\title{Function to create cell type annotation for Complex peak module heatmap}
\usage{
CreateCellTypeAnnotation(plot_data, scHelper_cell_type_colors)
}
\arguments{
\item{plot_data}{Made by PrepPeakModuleHeatmap function}

\item{scHelper_cell_type_colors}{vector of colors for scHelper_cell_type annotation}
}
\value{
Heatmap annotation for scHelper_cell_type
}
\description{
Function to create cell type annotation for Complex peak module heatmap
}
