% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStageAnnotation.R
\name{CreateStageAnnotation}
\alias{CreateStageAnnotation}
\title{Function to create stage annotation for Complex peak module heatmap}
\usage{
CreateStageAnnotation(plot_data, stage_colours)
}
\arguments{
\item{plot_data}{Made by PrepPeakModuleHeatmap function}

\item{scHelper_cell_type_colors}{vector of colors for stage annotation}
}
\value{
Heatmap annotation for stage
}
\description{
Function to create stage annotation for Complex peak module heatmap
}
