% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGeneModules.R
\name{DEGeneModules}
\alias{DEGeneModules}
\title{Filter gene modules based on differential expression (seurat::FindAllMarkers)}
\usage{
DEGeneModules(
  seurat_data,
  gene_modules,
  logfc = 0.25,
  pval = 0.001,
  selected_gene_proportion = 0.5,
  active_ident = NULL
)
}
\arguments{
\item{gene_modules}{named list of gene modules}

\item{logfc}{logFC cutoff used for differential expression}

\item{pval}{adjusted pvalue}

\item{selected_gene_proportion}{proportion of genes in gene module which are required to pass differential expression test}

\item{seurat_object}{Seurat object}
}
\value{
gene module list
}
\description{
Filter gene modules based on differential expression (seurat::FindAllMarkers)
}
