% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElbowCutoff.R
\name{ElbowCutoff}
\alias{ElbowCutoff}
\title{Elbow Cutoff}
\usage{
ElbowCutoff(seurat_object, return = "pc_cutoff")
}
\arguments{
\item{seurat_object}{Seurat object}

\item{return}{either 'pc_cutoff' or 'plot' based on whether the user would like to return a ggplot object or the principle component cutoff point}
}
\value{
ggplot object OR principle component cutoff
}
\description{
This function unbiasedly calculates where the principal components start to elbow.
}
\details{
First we take the larger value of the point where the principal components only contribute 5\% of standard deviation and the point where the principal components cumulatively contribute 90\% of the standard deviation.
Next we take the point where the percent change in variation between the consecutive PCs is less than 0.1\%.

The smaller out of these two values is determined at the elbow cutoff
}
