% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportAntlerModules.R
\name{ExportAntlerModules}
\alias{ExportAntlerModules}
\title{Exports antler peak modules}
\usage{
ExportAntlerModules(antler_object, publish_dir, names_list = "unbiasedPMs")
}
\arguments{
\item{antler_object}{antler object with the peak modules}

\item{publish_dir}{directory to write the peak modules to}

\item{names_list}{names of the peak modules to write out, as determined in antler object. Defaults to unbiasedPMs}
}
\value{
.txt files with the peak modules
}
\description{
This function writes out the peak modules calculated by antler to a text file
}
