% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractModuleContents.R
\name{ExtractModuleContents}
\alias{ExtractModuleContents}
\title{Extract module contents}
\usage{
ExtractModuleContents(modules_df, module_names)
}
\arguments{
\item{modules_df}{data frame of modules, each row name is module name and each column is a gene or peak ID}

\item{module_names}{vector of module names to extract, must match rownames of modules_df}
}
\value{
vector of peak or gene IDs
}
\description{
This function extracts the ids of peak or gene modules in a vector so they can be plotted
}
