% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModuleClassify.R
\name{GeneModuleClassify}
\alias{GeneModuleClassify}
\title{#' GMClassify}
\usage{
GeneModuleClassify(
  seurat_obj = seurat_data,
  gene_modules = antler_data$gene_modules$lists$unbiasedGMs_DE$content,
  metadata = NULL,
  order = NULL,
  rename_modules = FALSE,
  plot_path = "scHelper_log/GM_classification/"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene_modules}{List of gene modules generated by antler package}

\item{metadata}{metadata slot to group cells by e.g. stage, scHelper_cell_state, run}

\item{order}{vector of metadata values which determine how the gms should be ordered}

\item{plot_path}{directory to plot log files (boxplots)}
}
\value{
ordered dataframe of gene modules and their classifications
}
\description{
This function retrieves gene module classifications using cell module score columns in seurat object.
}
