% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModuleOrder.R
\name{GeneModuleOrder}
\alias{GeneModuleOrder}
\title{#' GMOrder}
\usage{
GeneModuleOrder(
  seurat_obj,
  gene_modules,
  metadata_1 = NULL,
  order_1 = NULL,
  metadata_2 = NULL,
  order_2 = NULL,
  rename_modules = NULL,
  plot_path = "scHelper_log/GM_classification/"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene_modules}{List of gene modules generated by antler package}

\item{metadata_1}{cell metadata column on which to order gms, first ordering}

\item{order_1}{vector specifying order of values from metadata_1}

\item{metadata_2}{cell metadata column on which to order gms, second ordering}

\item{order_2}{vector specifying order of values from metadata_2}

\item{rename_modules}{can be 'metadata_1' or 'metadata_2', will rename gms based on one of those classifications}

\item{plot_path}{Directory to plot log files from GMClassification}
}
\value{
list of gms ordered by expression levels in cells across metadata1 and 2
}
\description{
This function uses GMClassification to order GMs based on one or two cell metadata columns
}
