% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotColours.R
\name{ggPlotColours}
\alias{ggPlotColours}
\title{Function to access default ggpplot colours used by Seurat DimPlot}
\usage{
ggPlotColours(n = 6, h = c(0, 360) + 15)
}
\description{
Function to access default ggpplot colours used by Seurat DimPlot
}
