% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModuleClassification.R
\name{GMClassification}
\alias{GMClassification}
\title{#' GMClassification}
\usage{
GMClassification(
  seurat_obj = seurat_data,
  gene_modules = antler_data$gene_modules$lists$unbiasedGMs_DE$content,
  assay = "RNA",
  plot_path = "scHelper_log/GM_classification/",
  group_by = "scHelper_cell_type",
  publish_logs = TRUE,
  rotated_labels = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene_modules}{List of gene modules generated by antler package}

\item{plot_path}{Directory to plot log files}

\item{group_by}{column to group cells by, classification or clusters are recommended}

\item{publish_logs}{boolean value whether to publish logs. NOTE: setting to FALSE outputs a list of gm classifications and ggplot objects}

\item{rotated_labels}{boolean value to specify whether to rotate x axis labels in classification boxplots}
}
\value{
dataframe of gene modules and their classifications, unless publish_logs = FALSE
}
\description{
This function retrieves gene module classifications using cell module score columns in seurat object.
}
