% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyOutliers.R
\name{IdentifyOutliers}
\alias{IdentifyOutliers}
\title{Identify oulier clusters}
\usage{
IdentifyOutliers(
  dat,
  group_by = "seurat_clusters",
  metrics,
  intersect_metrics = FALSE
)
}
\arguments{
\item{dat}{df containing sample metadata (i.e. seurat@meta.data)}

\item{group_by}{column to group cells by}

\item{metrics}{array of metrics to test for outliers}

\item{intersect_metrics}{boolean value for whether to intersect the outliers for each metric}
}
\value{
array of clusters which are outliers
}
\description{
This function identifies clusters of cells based on QC metrics of interest. Groups are classed as outliers if the median for a given metric falls outside quartiles 1 and 3.
}
