% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFeaturePlot.R
\name{MultiFeaturePlot}
\alias{MultiFeaturePlot}
\title{Plot FeaturePlots for a list of genes}
\usage{
MultiFeaturePlot(
  seurat_object,
  gene_list,
  plot_stage = FALSE,
  stage_col = "orig.ident",
  plot_clusters = TRUE,
  cluster_col = "seurat_clusters",
  plot_celltype = FALSE,
  celltype_col = NA,
  n_col = 4,
  label = "UMAP plots for GOI on normalised filtered data"
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{gene_list}{genes to plot DimPlots for}

\item{plot_stage}{boolean for whether to plot DimPlot for developmental stage}

\item{stage_col}{seurat_object@meta.data column used to plot developmental stage}

\item{plot_clusters}{boolean for whether to plot DimPlot for clusters}

\item{cluster_col}{seurat_object@meta.data column used to plot clusters}

\item{plot_celltype}{boolean for whether to plot DimPlot for celltype}

\item{celltype_col}{seurat_object@meta.data column used to plot celltype}

\item{label}{plot title}
}
\value{
plot generated by grid extra
}
\description{
This function plots UMAPs for a series of nine cluster resolutions, alongside a clustree which reveals the relationship between clusters at different resolutions
}
