% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDistiller.R
\name{OptimizeRes}
\alias{OptimizeRes}
\title{自动化聚类寻优 (OptimizeRes)}
\usage{
OptimizeRes(
  object,
  res_range = seq(0.2, 1.2, 0.2),
  n_cells_sample = 5000,
  penalty_threshold = 4
)
}
\arguments{
\item{object}{Seurat 对象。必须已完成 \code{RunPCA}。}

\item{res_range}{数值向量。聚类分辨率的测试范围，默认 \code{seq(0.2, 1.2, 0.2)}。}

\item{n_cells_sample}{整数。用于计算轮廓系数的最大降采样细胞数（防内存溢出），默认 5000。}

\item{penalty_threshold}{整数。聚类数量过少时的惩罚阈值（鼓励中等粒度），默认 4。}
}
\value{
包含 \code{optimal_clusters} 新列与 \code{@misc[["scDistiller_best_res"]]} 记录的 Seurat 对象。
}
\description{
基于 PCA 降维结果，通过计算轮廓系数 (Silhouette Score) 自动寻找并推荐最佳聚类分辨率。
自动绘制不同分辨率下的平均轮廓系数折线图，并将最佳分辨率记录在对象的 misc 槽中。
}
\examples{
\dontrun{
pbmc <- OptimizeRes(pbmc, res_range = seq(0.2, 1.2, 0.2))
}
}
