% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderCellClusters.R
\name{OrderCellClusters}
\alias{OrderCellClusters}
\title{Automatic ordering of cell clusters based on cluster composition}
\usage{
OrderCellClusters(seurat_object, col_to_sort, sort_by)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{col_to_sort}{seurat_object@meta.data column to be re-arranged (i.e. clusters). Only rearrange cluster if at least 10\% of cells in cluster are from a different identity}

\item{sort_by}{array containing factor levels passed to col_to_sort (i.e. developmental stage)}
}
\value{
array of cluster IDs automatically ordered based on their composition
}
\description{
This function orders cell clusters by a column in the seurat metadata. i.e. sort seurat clusters by their developmental stage, with clusters shared between stages ordered in between those stages
}
