% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCALevelComparison.R
\name{PCALevelComparison}
\alias{PCALevelComparison}
\title{PCA level comparison}
\usage{
PCALevelComparison(data, PCA_levels = c(10, 15, 20, 30), cluster_res = 0.5)
}
\arguments{
\item{data}{Seurat object or list of Seurat objects}

\item{PCA_levels}{array of four max principle components to compare}

\item{cluster_res}{resolution used for KNN clustering}
}
\description{
Plot UMAPs using different numbers of principle components.
}
