% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCelltype.R
\name{PlotCelltype}
\alias{PlotCelltype}
\title{PlotCelltype}
\usage{
PlotCelltype(
  seurat_obj,
  gene_list,
  group_by = "seurat_clusters",
  x_lab = "Cluster ID",
  quantiles = c(0.05, 0.95),
  ...
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene_list}{Named list of genes passed to seurat::AddModuleScore}

\item{group_by}{seurat_object@meta.data column containing cell clusters}

\item{x_lab}{label for x axis}

\item{quantiles}{Percent quantile to plot on bar plots. Can be either a single element or an array (length 2) with lower and upper percent quantiles}

\item{...}{Extra arguments passed to grid.arrange}
}
\value{
plot output from grid.arrange
}
\description{
Calculate average expression for groups of genes across Seurat clusters and plot BoxPlots.
}
