% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneVariance.R
\name{PlotGeneVariance}
\alias{PlotGeneVariance}
\title{PlotGeneVariance}
\usage{
PlotGeneVariance(
  seurat_obj = seurat_data,
  group_by = "seurat_clusters",
  assay = "RNA"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{group_by}{column to group cells by}

\item{assay}{assay containing expression data. Default set to 'RNA'.}
}
\value{
plot
}
\description{
This function plots gene variance as a function of average expression for groups of cells.
}
