% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOutliers.R
\name{PlotOutliers}
\alias{PlotOutliers}
\title{PlotOutliers}
\usage{
PlotOutliers(
  seurat_obj,
  y_elements,
  group_by = "seurat_clusters",
  x_lab = "Cluster ID",
  quantiles
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{y_elements}{array of column names in seurat_object@meta.data to be plotted}

\item{group_by}{seurat_object@meta.data column containing cell clusters}

\item{x_lab}{label for x axis}

\item{quantiles}{Percent quantiles to plot on bar plots. Can be either NULL, a single element, or an array (length 2) with lower and upper percent quantiles}
}
\value{
list of box plots for each y_element
}
\description{
Plot box plots for elements in seurat_obj@meta.data with quartile cutoffs
}
