% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDistiller.R
\name{PlotTopMarkers}
\alias{PlotTopMarkers}
\title{极速特征提取与热图 (PlotTopMarkers)}
\usage{
PlotTopMarkers(object, cluster_col = "optimal_clusters", top_n = 5)
}
\arguments{
\item{object}{Seurat 对象。}

\item{cluster_col}{字符。用于提取 Marker 的聚类列名，默认 \code{"optimal_clusters"}。}

\item{top_n}{整数。每个 Cluster 提取并在热图中展示的 Top 基因数量，默认 5。}
}
\value{
一个包含 Top Marker 基因统计信息的 tibble 数据框。
}
\description{
利用 presto 包极速执行 Wilcoxon 秩和检验，提取各聚类群的 Top N 标志基因，
并自动绘制出版级标准化热图。
}
\examples{
\dontrun{
markers <- PlotTopMarkers(pbmc, cluster_col = "optimal_clusters", top_n = 5)
}
}
