% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepPeakModuleHeatmap.R
\name{PrepPeakModuleHeatmap}
\alias{PrepPeakModuleHeatmap}
\title{Function for prep data to plot peak module heatmap using Complex Heatmap}
\usage{
PrepPeakModuleHeatmap(
  peak_normalised_matrix,
  cell_metadata,
  col_order,
  custom_order_column = NULL,
  custom_order = NULL,
  hclust_SEACells = FALSE,
  hclust_SEACells_within_groups = TRUE,
  peak_modules,
  peak_row_annotation = TRUE,
  scale_data = TRUE,
  log_path = NULL
)
}
\arguments{
\item{peak_normalised_matrix}{matrix of peak normalised data, rows are peaks and columns are cells}

\item{cell_metadata}{data frame of cell metadata, rows are cells and columns are metadata}

\item{col_order}{vector of column names to order cells by}

\item{custom_order_column}{column name to order cells by custom ordering}

\item{custom_order}{vector of cell names to order cells by custom ordering}

\item{hclust_SEACells}{logical, whether to hclust SEACells}

\item{hclust_SEACells_within_groups}{logical, whether to hclust SEACells within SEACell groups}

\item{peak_modules}{list of peak modules, each element is a vector of peak IDs}

\item{peak_row_annotation}{logical, whether to add peak row annotation}

\item{scale_data}{logical, whether to scale data}

\item{log_path}{path to save logs (hclust dendograms), if NULL no logs will be saved}
}
\value{
plot_data list containing plot data, column annotation and row annotation
}
\description{
Function for prep data to plot peak module heatmap using Complex Heatmap
}
