% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDistiller.R
\name{PurifyCells}
\alias{PurifyCells}
\title{通用锚点过滤引擎 (PurifyCells)}
\usage{
PurifyCells(
  object,
  core_markers,
  whitelist_markers = NULL,
  expr_cutoff = 0.1,
  group_by = "condition",
  label_keep = "Kept",
  label_exclude = "Excluded",
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat 对象。}

\item{core_markers}{字符向量。核心目标群体表达的基因列表。}

\item{whitelist_markers}{字符向量。需要豁免保留的旁系细胞群基因列表，默认为 \code{NULL}。}

\item{expr_cutoff}{数值。判定的最低 log-normalized 表达量阈值，默认 0.1。}

\item{group_by}{字符。用于绘制纯化比例柱状图的分组列名，默认 \code{"condition"}。}

\item{label_keep}{字符。达标细胞的注释标签，默认 \code{"Kept"}。}

\item{label_exclude}{字符。未达标杂细胞的注释标签，默认 \code{"Excluded"}。}

\item{verbose}{逻辑值。是否在终端打印过滤统计信息，默认 \code{TRUE}。}
}
\value{
包含 \code{distilled_annotation} 新列的 Seurat 对象。
}
\description{
为特定下游分析（如 CellChat / NicheNet）精准切取目标细胞生态位。
基于给定的核心基因和白名单基因表达阈值，将数据集二元划分为保留群体与背景群体。
}
\examples{
\dontrun{
pbmc <- PurifyCells(pbmc, core_markers = c("CD3D"), whitelist_markers = c("CD14"))
}
}
