% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCPlot.R
\name{QCPlot}
\alias{QCPlot}
\title{Generate a QC plot from Seurat metadata}
\usage{
QCPlot(
  seurat_obj,
  group_by = "seurat_clusters",
  y_elements = c("nCount_RNA", "nFeature_RNA", "percent.mt"),
  y_lab = c("UMI Count", "Gene Count", "\% MT"),
  x_lab = "Cluster ID"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{group_by}{seurat_obj@meta.data column to group cells by on x axis}

\item{y_elements}{seurat_obj@meta.data column to plot QC metrics for}

\item{y_lab}{array of labels for y axis reflecting y_elements}

\item{x_lab}{label for x axis}
}
\value{
QC plots
}
\description{
This function generates a series of box plots as quality control for a Seurat object
}
