% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHelperClass.R
\name{scHelper-class}
\alias{scHelper-class}
\alias{scHelper}
\title{An S4 class to extend Seurat class to include scHelper specific slots.}
\usage{
scHelper(...)
}
\description{
An S4 class to extend Seurat class to include scHelper specific slots.
}
\section{Slots}{

\describe{
\item{\code{feature.meta}}{A dataframe containing feature information. At minimum containing gene identifiers and associated gene names}

\item{\code{gene.modules}}{A list of gene modules obtained from Antler}
}}

