% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACells_MetacellFrequencies.R
\name{SEACells_MetacellFrequencies}
\alias{SEACells_MetacellFrequencies}
\title{SEACells function: Frequency of metacells in each category}
\usage{
SEACells_MetacellFrequencies(
  seurat,
  metacell_slot = "SEACell",
  category = "stage",
  calc_proportions = FALSE
)
}
\arguments{
\item{seurat}{seurat object}

\item{metacell_slot}{slot in seurat metadata by which to group cells for summarising, default is "SEACell"}

\item{category}{slot in seurat metadata by which to count metacell frequencies, default is "stage"}

\item{calc_proportions}{boolean as to whether to calculate proportions of metacells in each category}
}
\value{
dataframe of either frequency of metacells in each category or proportions of metacells in each category
}
\description{
Function to take seurat object and a category and make a freq table of how frequently metacells found in each category
}
