% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACells_PiechartProportionThreshold.R
\name{SEACells_PiechartProportionThreshold}
\alias{SEACells_PiechartProportionThreshold}
\title{SEACells function: Purity of metacells plotted on piechart}
\usage{
SEACells_PiechartProportionThreshold(prop_table, threshold = 0.5)
}
\arguments{
\item{prop_table}{dataframe of proportions of metacells in each category, output from scHelper::calculate_metacell_frequencies}

\item{threshold}{threshold for proportion of cells coming from one label}
}
\value{
pie chart showing proportions of cells that pass threshold purity
}
\description{
Function to plot piechart of how many metacells pass threshold for proportion of cells coming from one label
}
