% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEACells_SummariseGeneScoreMatrix.R
\name{SEACells_SummariseGeneScoreMatrix}
\alias{SEACells_SummariseGeneScoreMatrix}
\title{SEACells function: Summarises gene score counts across SEACells}
\usage{
SEACells_SummariseGeneScoreMatrix(matrix, dictionary)
}
\arguments{
\item{matrix}{exported gene score matrix from SEACells}

\item{dictionary}{dataframe with two columns: 'index' (single cell ID) and 'SEACell' (SEACell ID)}
}
\value{
dataframe with summarised gene counts which can be used to make a seurat object of SEACells
}
\description{
This function takes ATAC gene score matrix and adds up gene score counts across SEACells
}
