% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHelper.R
\name{shFilter}
\alias{shFilter}
\title{Step 9: 目标亚群切取 (shFilter)}
\usage{
shFilter(
  object,
  core_markers,
  whitelist_markers = NULL,
  expr_cutoff = 0.1,
  group_by = "condition",
  label_keep = "Kept",
  label_exclude = "Excluded"
)
}
\arguments{
\item{object}{Seurat 对象。}

\item{core_markers}{字符向量。核心需要保留的亚群特征基因。}

\item{whitelist_markers}{字符向量。额外需要保留的亚群特征基因（可选）。}

\item{expr_cutoff}{数值。基因表达量判定阈值，默认 0.1。}

\item{group_by}{字符。用于统计提纯比例的分组列，默认 "condition"。}

\item{label_keep}{字符。保留簇标签名，默认 "Kept"。}

\item{label_exclude}{字符。排除簇标签名，默认 "Excluded"。}
}
\value{
新增 \code{distilled_annotation} 二分类标签列的 Seurat 对象。
}
\description{
基于核心和白名单 Marker 基因列表，精准切取所需亚群并标记保留状态。
}
