% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHelper.R
\name{shHVG}
\alias{shHVG}
\title{Step 3: HVG 数量寻优 (shHVG)}
\usage{
shHVG(object, max_features = 5000)
}
\arguments{
\item{object}{Seurat 对象。}

\item{max_features}{整数。评估的最大特征数，默认 5000。}
}
\value{
完成 FindVariableFeatures 并记录最优点至 \code{scHelper_optimal_hvg} 的 Seurat 对象。
}
\description{
利用累积标准化方差饱和曲线与几何弦距法自动选择最优高变基因数量。
}
