% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHelper.R
\name{shPC}
\alias{shPC}
\title{Step 4: 主成分数量寻优 (shPC)}
\usage{
shPC(object, max_pc = 50)
}
\arguments{
\item{object}{Seurat 对象。必须已完成 RunPCA。}

\item{max_pc}{整数。评估的最大 PC 数，默认 50。}
}
\value{
记录最优点至 \code{scHelper_optimal_pc} 的 Seurat 对象。
}
\description{
利用几何弦距法在主成分方差碎石图上自动定位数学拐点。
}
