% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_SplitbyStage.R
\name{split_ArchR_by_stage}
\alias{split_ArchR_by_stage}
\title{ArchR function: Split by stage}
\usage{
split_ArchR_by_stage(ArchR)
}
\arguments{
\item{ArchR}{ArchR object}
}
\value{
list of ArchR objects, split by stage
}
\description{
This function splits ArchR object by stage
}
