% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetGeneModules.R
\name{SubsetGeneModules}
\alias{SubsetGeneModules}
\title{Subset Antler gene modules based on selected gene list provided}
\usage{
SubsetGeneModules(
  gms,
  selected_genes,
  keep_mod_ID = FALSE,
  selected_gene_proportion = 0
)
}
\arguments{
\item{gms}{List of gene modules obtained from Antler}

\item{selected_genes}{array of genes used to filter gene module list}

\item{keep_mod_ID}{boolean value for whether to keep original gene module ID}

\item{selected_gene_proportion}{ratio of selected genes which must be present in order for the module to be kept}
}
\description{
Subset Antler gene modules based on selected gene list provided
}
