% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TenxPheatmap.R
\name{TenxPheatmap}
\alias{TenxPheatmap}
\title{Function for plotting heatmap from seurat 10x object}
\usage{
TenxPheatmap(
  data,
  metadata,
  col_order = metadata,
  custom_order = NULL,
  custom_order_column = NULL,
  assay = "RNA",
  slot = "scale.data",
  selected_genes,
  main = "",
  hide_annotation = NULL,
  show_rownames = TRUE,
  hclust_rows = FALSE,
  hclust_cols = FALSE,
  gaps_col = NULL,
  cell_subset = NULL,
  treeheight_row = 0,
  use_seurat_colours = TRUE,
  colour_scheme = c("PRGn", "RdYlBu", "Greys"),
  col_ann_order = rev(metadata)
)
}
\description{
col_order is a single element or a vector element. By default cells are ordered by the first element of the metadata variable.
User can specify col_order as a vector, with cell order prioritised based on the order of the col_order variable.
}
\details{
A custom_order can be provided, however if so the custom_order_column must also be specified.

By default cells are coloured based on the default ggplot colours (to match seurat cluster colours).
This behaviour can be changed by changing use_seurat_colours and colour_scheme (RColourBrewer)

col_ann_order can be specified to change the order in which the column annotations appear on the heatmap
}
