% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAPGeneList.R
\name{UMAPGeneList}
\alias{UMAPGeneList}
\title{Generate individual FeaturePlots for given list of genes}
\usage{
UMAPGeneList(seurat_object, gene_list, plot_path, zip_files = TRUE)
}
\arguments{
\item{seurat_object}{seurat object}

\item{gene_list}{array of genes to plot}

\item{plot_path}{path to plot FeaturePlots to}

\item{zip_files}{boolean for whether or not to zip plots}
}
\description{
Generate individual FeaturePlots for given list of genes
}
