% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterClassification.R
\name{ClusterClassification}
\alias{ClusterClassification}
\title{ClusterClassification}
\usage{
ClusterClassification(
  seurat_obj = seurat_data,
  group_by = "seurat_clusters",
  cell_type_markers = cell_type_markers,
  quantile = 0.8,
  assay = "RNA",
  annotation_name = "scHelper_cell_type",
  force_assign = FALSE,
  plot_path = "scHelper_log/",
  ...
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{group_by}{column to group cells by}

\item{quantile}{Percent quantile to plot on bar plots. Must be a single value representing the upper percent quantile}

\item{annotation_name}{Name of column to be annotated with corresponding cell types}

\item{force_assign}{Boolean for whether to force assign cell identity to clusters which do not pass the filtering threshold}

\item{plot_path}{Directory to plot log files}

\item{...}{Additional options to be passed to grid.arrange}

\item{cell_types}{column names for cell type module scores found in seurat_obj@meta.data}
}
\value{
array of clusters which are outliers
}
\description{
This function retrieves cluster classifications using cell module score columns in seurat object. Clusters are classiffied as a given cell type if the median for a given cell type falls outside the provided quantile.
}
