% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustStagePlot.R
\name{ClustStagePlot}
\alias{ClustStagePlot}
\title{DimPlots of cell stage and cluster identities}
\usage{
ClustStagePlot(
  seurat_object,
  cluster_col = "seurat_clusters",
  stage_col = "orig.ident"
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{cluster_col}{seurat_object@meta.data column containing cell clusters}

\item{stage_col}{seurat_object@meta.data column containing cell stage}
}
\value{
plot generated by grid extra
}
\description{
This function plots UMAPs for cell stage and cluster identities
}
